Move premises utilities out of service
