+++ server/views/temporary-accommodation/bedspaces/show.njk
@@ -43,7 +43,7 @@
             <h2 class="govuk-heading-s">Property address</h2>
 
             <p class="govuk-body">
-                {{ premises.fullAddress | safe }}
+                {{ fullAddress | safe }}
             </p>
         </div>
